:sparkles: Add graceful server shutdown and logging
A graceful server shutdown functionality has been introduced. This allows the server to shutdown properly when signals such as SIGINT and SIGTERM are received. It includes a shutdown hook, which is configured to stop the server, and incorporated relevant error logs in case of a shutdown error. Additionally, few modifications are made in MongoDB's Close connection and a log message indicating server shutdown is added to the `OnShutdown` hook.